Refactor out summary components
