+++ src/utility.ts
@@ -1,4 +1,4 @@
-import { Comment, Post, RedditAPIClient } from "@devvit/public-api";
+import { Comment, Post, RedditAPIClient, TriggerContext } from "@devvit/public-api";
 import { isCommentId, isLinkId, isSubredditId, T1ID, T3ID, T5ID } from "@devvit/shared-types/tid.js";
 
 export async function getPostOrCommentFromRedditId (reddit: RedditAPIClient, redditId?: T1ID | T3ID | T5ID): Promise <Post | Comment | undefined> {
@@ -9,4 +9,13 @@ export async function getPostOrCommentFromRedditId (reddit: RedditAPIClient, red
     } else if (isLinkId(redditId)) {
         return reddit.getPostById(redditId);
     }
+}
+
+export async function getSubredditName (context: TriggerContext): Promise<string> {
+    if (context.subredditName) {
+        return context.subredditName;
+    }
+
+    // This shouldn't happen, but add a fallback just in case.
+    return (await context.reddit.getCurrentSubreddit()).name;
 }
